+++ CHANGES.md
@@ -10,3 +10,4 @@
 - Improved Pure Love visuals
 - Nerfed Output RCT
 - Slicing entities with dismantle
+- GeckoLib model support for dismantle